requirements.txt
streamlit
requests
pandas
st-annotated-text
python-dotenv